+++ app/models/mokio/seo_tag.rb
@@ -1,0 +1,5 @@
+module Mokio
+  class SeoTag < ActiveRecord::Base
+    include Mokio::Concerns::Models::SeoTag
+  end
+end